file name converted to .jsx
